chore(administration): split models
