+++ public_data/models/administration/Scot.py
@@ -1,0 +1,54 @@
+from django.contrib.gis.db import models
+
+from public_data.models.cerema import Cerema
+from public_data.models.enums import SRID
+from utils.db import IntersectManager
+
+from .AdminRef import AdminRef
+from .GetDataFromCeremaMixin import GetDataFromCeremaMixin
+from .LandMixin import LandMixin
+
+
+class Scot(LandMixin, GetDataFromCeremaMixin, models.Model):
+    name = models.CharField("Nom", max_length=250)
+    mpoly = models.MultiPolygonField(srid=4326, null=True, blank=True)
+    srid_source = models.IntegerField(
+        "SRID",
+        choices=SRID.choices,
+        default=SRID.LAMBERT_93,
+    )
+
+    regions = models.ManyToManyField("Region")
+    departements = models.ManyToManyField("Departement")
+    siren = models.CharField("Siren", max_length=12, null=True, blank=True)
+
+    objects = IntersectManager()
+
+    land_type = AdminRef.SCOT
+    default_analysis_level = AdminRef.EPCI
+
+    @property
+    def official_id(self) -> str:
+        return self.siren
+
+    def get_qs_cerema(self):
+        return Cerema.objects.filter(city_insee__in=self.commune_set.values("insee"))
+
+    def get_cities(self):
+        return self.commune_set.all()
+
+    def __str__(self):
+        return self.name.upper()
+
+    def get_official_id(self) -> str:
+        return self.siren if self.siren is not None else ""
+
+    @classmethod
+    def search(cls, needle, region=None, departement=None, epci=None):
+        qs = cls.objects.filter(name__unaccent__trigram_word_similar=needle)
+        if region:
+            qs = qs.filter(regions=region)
+        if departement:
+            qs = qs.filter(id=departement.id)
+        qs = qs.order_by("name")
+        return qs
